[refactoring] 라이딩 일지 컴포넌트 분리
