+++ src/components/Users/RidingRecord/items/List/RecordListTitle.js
@@ -1,0 +1,16 @@
+import React from "react";
+
+const RecordListTitle = () => {
+  return (
+    <ul className="record-title">
+      <li>날짜</li>
+      <li>제목</li>
+      <li>거리</li>
+      <li>시간</li>
+      <li>라이딩 점수</li>
+      <li>&nbsp;</li>
+    </ul>
+  );
+};
+
+export default RecordListTitle;
