+++ src/components/Users/RidingRecord/pages/Main.js
@@ -1,11 +1,12 @@
 import React, { useState, useEffect } from "react";
 import Title from "../../../item/Title";
-import WeekStat from "../items/WeekStat";
 import "./Main.css";
 
-import Dropdown from "react-dropdown";
 import "react-dropdown/style.css";
 import { getRidingRecordByYear } from "../../../../api/RidingRecord";
+import RecordMainHeader from "../items/Main/RecordMainHeader";
+import RecordMainChartTitle from "../items/Main/RecordMainChartTitle";
+import RecordMainChartContens from "../items/Main/RecordMainChartContens";
 
 const Main = () => {
   const currentYear = new Date().getFullYear();
@@ -45,49 +46,20 @@ const Main = () => {
     setPrevYear(year);
     setYear(value);
   };
+  const header = {
+    years,
+    year,
+    yearChangeHandler,
+  };
 
   return (
     <>
       <Title title="라이딩 일지" />
       <div className="record-main">
-        <div className="header">
-          <div className="dropdown-section">
-            <Dropdown
-              options={years}
-              onChange={yearChangeHandler}
-              value={year}
-              placeholder="조회년도"
-            />
-          </div>
-          <div className="dropdown-txt">년</div>
-          <div className="legend-section">
-            <div>거리</div>
-            <div>시간</div>
-            <div>평균속도</div>
-          </div>
-        </div>
+        <RecordMainHeader data={header} />
         <div className="chart-section">
-          <ul className="chart-title">
-            <li>{year}</li>
-            <li>일</li>
-            <li>월</li>
-            <li>화</li>
-            <li>수</li>
-            <li>목</li>
-            <li>금</li>
-            <li>토</li>
-          </ul>
-          <ul className="chart-contents">
-            {/* FIXME stats object -> array 로 변경 필요, 변경 시 바로 map 적용 */}
-            {Object.values(stats)
-              .reverse()
-              .map((stat, idx) => {
-                return <WeekStat key={idx} stat={stat} year={year} />;
-              })}
-            {/* {stats.map((stat, idx) => {
-              return <WeekStat key={idx} stat={stat} year={year} />;
-            })} */}
-          </ul>
+          <RecordMainChartTitle year={year} />
+          <RecordMainChartContens year={year} stats={stats} />
         </div>
       </div>
     </>
